requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
google-api-core==1.22.2
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.28.0
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
